Moved Data, New Animation code from Github
I moved the data files to a dedicated folder to reduce clutter. We can organize this further as needed.

I also implemented a neat bit of code, see Jake_animate.py for original codes. I modified it to make a kind of movie of the difference between a centre-cropped image and small shifts in that crop position.